requirements.txt
torch==2.5.1
numpy==2.1.3
transformers==4.46.2
wandb==0.18.7
datasets==3.1.0
tqdm==4.67.0
torchvision==0.19.0
pillow==10.3.0
decord==0.6.0
timm==0.9.16